requirements.txt
# Web scraping
selenium>=4.15.0

# PDF text extraction
PyMuPDF>=1.23.0

# Web UI backend
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# AI summarization
groq>=0.12.0

# Environment variable management
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0

# Linting (optional, nice for resume projects)
ruff>=0.1.0